requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-rest-registration==0.5.6
django-simple-history==2.11.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
httpie==2.2.0
idna==2.10
psycopg2-binary==2.8.6
Pygments==2.7.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.3
urllib3==1.25.10
django-environ-0.4.5